added marker on the map corresponded to UAV ID as user add
