--- flight-visualizer-js/src/components/UAV/UAV.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react'
-
-class UAV extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {}
-  }
-
-  updateMarkerPosition() {}
-
-  render() {
-    return <div></div>
-  }
-}
-
-export default UAV
